Fixed an issue with the activity dialogue not showing the decimal value of the price of a grocery item

+++ app/src/main/java/cse5321/roommateapp/Grocery.java
@@ -153,7 +153,7 @@ public class Grocery extends ParseObject {
      * @return the price the grocery
      */
     public double getPrice() {
-        return getInt("Price");
+        return getDouble("Price");
     }
 
     /**
